# docs/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "EduMind MVP 0.3 API contract",
    "version": "0.3.0",
    "description": "MVP 0.1/0.2已实现接口及MVP 0.3拟实现契约；新增路径仅为T004设计，运行端点由T012–T028交付。JSON是YAML 1.2子集。"
  },
  "servers": [{"url": "/"}],
...
  "paths": {
...
    "/api/learning-units/{learningUnitId}/animations": {"post": {"tags": ["Animation"],"summary": "按需请求可信链表动画","operationId": "requestAnimation","security": [{"sessionCookie": []}],"parameters": [{"$ref": "#/components/parameters/LearningUnitId"},{"$ref": "#/components/parameters/CsrfToken"},{"$ref": "#/components/parameters/IdempotencyKey"}],"responses": {"202": {"description": "成功。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/AnimationJob"}}}},"200": {"description": "缓存命中或幂等重放。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/AnimationJob"}}}},"401": {"$ref": "#/components/responses/Unauthorized"},"403": {"$ref": "#/components/responses/CsrfFailed"},"404": {"$ref": "#/components/responses/NotFound"},"409": {"$ref": "#/components/responses/IdempotencyConflict"},"422": {"$ref": "#/components/responses/ValidationError"},"503": {"$ref": "#/components/responses/RenderUnavailable"}},"requestBody": {"required": true,"content": {"application/json": {"schema": {"$ref": "#/components/schemas/AnimationRequest"}}}},"description": "owner+操作类别+幂等键+摘要；同键同摘要回放，不重新渲染。当前P0教学场景选用学习单元intro的已审核版本，知识点须与模板ID一致。缓存命中仍创建owner授权绑定。"}},
    "/api/animation-jobs/{jobId}": {"get": {"tags": ["Animation"],"summary": "读取Job快照及事件水位","operationId": "getAnimationJob","security": [{"sessionCookie": []}],"parameters": [{"$ref": "#/components/parameters/AnimationJobId"}],"responses": {"200": {"description": "成功。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/AnimationJob"}}}},"401": {"$ref": "#/components/responses/Unauthorized"},"404": {"$ref": "#/components/responses/NotFound"}}}},
    "/api/animation-jobs/{jobId}/cancel": {"post": {"tags": ["Animation"],"summary": "取消未完成Job","operationId": "cancelAnimationJob","security": [{"sessionCookie": []}],"parameters": [{"$ref": "#/components/parameters/AnimationJobId"},{"$ref": "#/components/parameters/CsrfToken"},{"$ref": "#/components/parameters/IdempotencyKey"}],"responses": {"200": {"description": "成功。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/AnimationJob"}}}},"401": {"$ref": "#/components/responses/Unauthorized"},"403": {"$ref": "#/components/responses/CsrfFailed"},"404": {"$ref": "#/components/responses/NotFound"},"409": {"description": "IDEMPOTENCY_CONFLICT 或 当前Job终态不允许该操作；code=JOB_STATE_CONFLICT。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},"422": {"$ref": "#/components/responses/ValidationError"}},"description": "已成功终态不能取消；取消幂等键按owner唯一持久保存，取消与发布由同一Job行锁确定顺序。Worker检测到取消后终止当前尝试容器。"}},
    "/api/animation-jobs/{jobId}/retry": {"post": {"tags": ["Animation"],"summary": "显式重试失败或已取消Job","operationId": "retryAnimationJob","security": [{"sessionCookie": []}],"parameters": [{"$ref": "#/components/parameters/AnimationJobId"},{"$ref": "#/components/parameters/CsrfToken"},{"$ref": "#/components/parameters/IdempotencyKey"}],"responses": {"202": {"description": "成功。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/AnimationJob"}}}},"401": {"$ref": "#/components/responses/Unauthorized"},"403": {"$ref": "#/components/responses/CsrfFailed"},"404": {"$ref": "#/components/responses/NotFound"},"409": {"description": "IDEMPOTENCY_CONFLICT 或 当前Job终态不允许该操作；code=JOB_STATE_CONFLICT。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},"422": {"$ref": "#/components/responses/ValidationError"}},"description": "创建新Job并保存retry_of；原终态不可修改。"}},
    "/api/animation-jobs/{jobId}/events": {"get": {"tags": ["Animation"],"summary": "持久Job事件流","operationId": "subscribeAnimationJob","security": [{"sessionCookie": []}],"parameters": [{"$ref": "#/components/parameters/AnimationJobId"},{"$ref": "#/components/parameters/LastEventId"},{"$ref": "#/components/parameters/AnimationEventAfter"}],"responses": {"200": {"description": "GET EventSource text/event-stream；事件ID为Job内单调整数。","content": {"text/event-stream": {"schema": {"type": "string"}}}},"401": {"$ref": "#/components/responses/Unauthorized"},"404": {"$ref": "#/components/responses/NotFound"},"409": {"$ref": "#/components/responses/EventCursorInvalid"},"410": {"$ref": "#/components/responses/EventCursorExpired"},"503": {"description": "本进程SSE连接容量已满；code=STREAM_CAPACITY。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}},"description": "无游标从头；Last-Event-ID只回放更大ID。终态且游标已消费则立即关闭。410后GET快照取last_event_id再订阅；心跳无持久ID。每进程最多32连接，轮询间隔1.5秒，单连接最多5分钟后由EventSource按游标重连；首次连接可用after查询参数，重连时Last-Event-ID头优先。"}},
...
  },
  "components": {
    "securitySchemes": {"sessionCookie": {"type": "apiKey", "in": "cookie", "name": "edumind_session", "description": "Secure、HttpOnly、SameSite=Lax、Host-only Cookie，2h 绝对有效期；数据库只存令牌哈希。"}},
    "parameters": {
      "CsrfToken": {"name": "X-CSRF-Token", "in": "header", "required": true, "schema": {"type": "string", "minLength": 64, "maxLength": 64}, "description": "写请求必须提供；服务端同时验证同源 Origin。"},
      "IdempotencyKey": {"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"type": "string", "minLength": 16, "maxLength": 128}, "description": "客户端为一次创建操作生成的随机键；与认证身份和请求摘要绑定。"},
...
      "LearningUnitId": {"name": "learningUnitId", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/Identifier"}},
...
      "AnimationJobId": {"name": "jobId","in": "path","required": true,"schema": {"$ref": "#/components/schemas/Identifier"}},
...
      "LastEventId": {"name": "Last-Event-ID","in": "header","required": false,"schema": {"type": "integer","minimum": 0},"description": "动画Job事件：仅重放ID更大的已提交事件；过期410后GET快照取得last_event_id。"},
      "AnimationEventAfter": {"name": "after","in": "query","required": false,"schema": {"type": "integer","minimum": 0,"maximum": 999999999999999999},"description": "EventSource首次订阅不能自定义请求头，可用快照last_event_id作为初始游标；自动重连的Last-Event-ID头优先于本参数。"},
...
    },
    "responses": {
...
      "Unauthorized": {"description": "会话缺失、失效或不属于当前用户。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "CsrfFailed": {"description": "Origin 或 CSRF token 校验失败。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "NotFound": {"description": "资源不存在，或不属于当前用户。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "ValidationError": {"description": "请求未通过 schema 校验。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "IdempotencyConflict": {"description": "同一幂等键被用于不同请求摘要。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
      "EventCursorInvalid": {"description": "动画事件游标超过当前最大值；code=EVENT_CURSOR_INVALID。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "EventCursorExpired": {"description": "动画事件已清理；code=EVENT_CURSOR_EXPIRED，GET快照取得last_event_id后重订阅。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
      "RenderUnavailable": {"description": "隔离执行器不可用或无法建立限制；code=RENDER_UNAVAILABLE。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
    },
    "schemas": {
...
      "Identifier": {"type": "string", "pattern": "^[A-Za-z0-9_-]{8,64}$"},
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "UNAUTHORIZED",
              "CSRF_FAILED",
              "NOT_FOUND",
              "VALIDATION_ERROR",
              "IDEMPOTENCY_CONFLICT",
              "PROFILE_VERSION_CONFLICT",
              "PATH_VERSION_CONFLICT",
              "GRAPH_NODE_NOT_FOUND",
              "QUIZ_SUBMISSION_INVALID",
              "PATH_UNREACHABLE",
              "ACCOUNT_CONFLICT",
              "RATE_LIMITED",
              "PROVIDER_UNAVAILABLE",
              "REVIEW_REJECTED",
              "OPERATION_CANCELED",
              "JOB_STATE_CONFLICT",
              "RENDER_UNAVAILABLE",
              "RENDER_FAILED",
              "RENDER_INTERRUPTED",
              "MEDIA_UNAVAILABLE",
              "MEDIA_NOT_READY",
              "CLASSROOM_VERSION_CONFLICT",
              "SCENE_VERSION_CONFLICT",
              "EVENT_CURSOR_INVALID",
              "EVENT_CURSOR_EXPIRED",
              "MESSAGE_CURSOR_INVALID",
              "MESSAGE_CURSOR_EXPIRED",
              "REVIEW_UNAVAILABLE",
              "CONTENT_RETRACTED",
              "ROLE_DISABLED",
              "DEBATE_UNAVAILABLE",
              "DEBATE_ALREADY_ACTIVE",
              "DEBATE_NOT_ACTIVE",
              "DEBATE_CONTEXT_UNAVAILABLE",
              "DEBATE_PUBLICATION_FAILED",
              "DEBATE_INVALID_OUTPUT",
              "DEBATE_INTERRUPTED"
            ]
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "operation_id": {
            "$ref": "#/components/schemas/Identifier"
          }
        }
      },
      "AnimationRequest": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/AnimationInsertRequest"
          },
          {
            "$ref": "#/components/schemas/AnimationDeleteRequest"
          }
        ],
        "description": "索引与当前values长度的关系由服务端固定模板注册表进一步校验；owner学习单元知识点必须与template_id相同。"
      },
      "AnimationJob": {"type": "object","additionalProperties": false,"properties": {"id": {"$ref": "#/components/schemas/Identifier"},"status": {"type": "string","enum": ["queued","running","succeeded","failed","cancelled"]},"learning_unit_id": {"$ref": "#/components/schemas/Identifier"},"template_id": {"type": "string"},"attempt": {"type": "integer","minimum": 0},"progress": {"type": "number","minimum": 0,"maximum": 1},"last_event_id": {"type": "integer","minimum": 0},"media_id": {"$ref": "#/components/schemas/Identifier"},"retry_of": {"$ref": "#/components/schemas/Identifier"},"error": {"$ref": "#/components/schemas/Error"}},"required": ["id","status","learning_unit_id","template_id","attempt","progress","last_event_id"]},
...
      "AnimationInsertRequest": {"type": "object","additionalProperties": false,"required": ["template_id","template_version","scene_version","parameters"],"properties": {"template_id": {"const": "linked-list-insertion"},"template_version": {"type": "string","pattern": "^[0-9]+\\.[0-9]+\\.[0-9]+$","maxLength": 40},"scene_version": {"type": "integer","minimum": 1},"parameters": {"type": "object","additionalProperties": false,"required": ["values","index","value"],"properties": {"values": {"type": "array","minItems": 0,"maxItems": 7,"items": {"type": "integer","minimum": -99,"maximum": 99}},"index": {"type": "integer","minimum": 0,"maximum": 7},"value": {"type": "integer","minimum": -99,"maximum": 99}}}}},
      "AnimationDeleteRequest": {"type": "object","additionalProperties": false,"required": ["template_id","template_version","scene_version","parameters"],"properties": {"template_id": {"const": "linked-list-deletion"},"template_version": {"type": "string","pattern": "^[0-9]+\\.[0-9]+\\.[0-9]+$","maxLength": 40},"scene_version": {"type": "integer","minimum": 1},"parameters": {"type": "object","additionalProperties": false,"required": ["values","index"],"properties": {"values": {"type": "array","minItems": 1,"maxItems": 8,"items": {"type": "integer","minimum": -99,"maximum": 99}},"index": {"type": "integer","minimum": 0,"maximum": 7}}}}}
    }
  },
...
}